requirements.txt
python-dotenv
Pillow
google-generativeai
pandas
pyarrow